python-backend/entangl-fact-checker/requirements.txt
requests
feedparser
python-dotenv
google-generativeai
